Initial Commit
Completed start to finish in 14 hours.  This is a basic php based
e-commerce site with basic MySQL backend, with jQuery and Materialize
CSS/JS on the front end.

Disclaimer: I do not own or claim the rights to any images included in
this project.  The images used in this project are used only as
examples.  Links to their respective sources are included in the source
code.